v'2.0'

+++ deletelist.php
@@ -1,20 +1,35 @@
 <?php
 include("header.php");
-$id=$_GET["id"];
 
+$id=$_GET["id"];
+$fileurl="";
+
+$link=mysqli_connect("localhost","root","","fastfood shop");
+$result=mysqli_query($link,"SELECT * FROM `mahsolat` WHERE id=$id");
+mysqli_close($link);
+$row=mysqli_fetch_array($result);
+if($row){
+    $fileurl=$row["imageurl"];
+}
+
+$delete=unlink($fileurl);
+if($delete==true){
 $link=mysqli_connect("localhost","root","","fastfood shop");
 $result=mysqli_query($link,"DELETE FROM `mahsolat` WHERE `id`=$id;");
 mysqli_close($link);
 
-?>
-<div class="row">
-    <div class="col">
+if($result===true)
+{
+    ?>
     <script>
         location.replace("list.php");
     </script>
-    </div>
-</div>
+    <?php
+}else
+{
+    echo("عملیات با مشکل مواجه شد ❌");
+}
+}
 
-<?php
 include("footer.html");
 ?>
